ACTUAL FIX: Remove public/ from .gitignore
The public/ folder was being ignored by Git (Gatsby leftover)
This meant NO images were ever committed to the repository!

Astro requires the public/ folder to be committed because it
contains static assets that should be deployed.

This was the root cause - images never made it to Netlify because
they were never in Git in the first place.